de-clutter transaction.py
remove unused functions and move opcodes, Enumeration, and BCDataStream
to their own files
