+++ lbryum/constants.py
@@ -14,6 +14,8 @@ TYPE_UPDATE = 32
 EXPIRATION_BLOCKS = 262974
 RECOMMENDED_CLAIMTRIE_HASH_CONFIRMS = 1
 
+NO_SIGNATURE = 'ff'
+
 NULL_HASH = '0000000000000000000000000000000000000000000000000000000000000000'
 HEADER_SIZE = 112
 BLOCKS_PER_CHUNK = 96
